+++ packages/models/src/Domain/Syncable/UserPrefs/PrefKey.ts
@@ -9,6 +9,8 @@ export enum PrefKey {
   EditorMonospaceEnabled = 'monospaceFont',
   EditorSpellcheck = 'spellcheck',
   EditorResizersEnabled = 'marginResizersEnabled',
+  EditorLineHeight = 'editorLineHeight',
+  EditorFontSize = 'editorFontSize',
   SortNotesBy = 'sortBy',
   SortNotesReverse = 'sortReverse',
   NotesShowArchived = 'showArchived',
@@ -33,12 +35,31 @@ export enum PrefKey {
   MobileSelectedTagUuid = 'mobileSelectedTagUuid',
   MobileNotesHideEditorIcon = 'mobileHideEditorIcon',
   NewNoteTitleFormat = 'newNoteTitleFormat',
+  CustomNoteTitleFormat = 'customNoteTitleFormat',
 }
 
 export enum NewNoteTitleFormat {
   CurrentDateAndTime = 'CurrentDateAndTime',
   CurrentNoteCount = 'CurrentNoteCount',
+  CustomFormat = 'CustomFormat',
   Empty = 'Empty',
+}
+
+export enum EditorLineHeight {
+  None = 'None',
+  Tight = 'Tight',
+  Snug = 'Snug',
+  Normal = 'Normal',
+  Relaxed = 'Relaxed',
+  Loose = 'Loose',
+}
+
+export enum EditorFontSize {
+  ExtraSmall = 'ExtraSmall',
+  Small = 'Small',
+  Normal = 'Normal',
+  Medium = 'Medium',
+  Large = 'Large',
 }
 
 export type PrefValue = {
@@ -73,4 +94,7 @@ export type PrefValue = {
   [PrefKey.MobileSelectedTagUuid]: string | undefined
   [PrefKey.MobileNotesHideEditorIcon]: boolean
   [PrefKey.NewNoteTitleFormat]: NewNoteTitleFormat
+  [PrefKey.CustomNoteTitleFormat]: string
+  [PrefKey.EditorLineHeight]: EditorLineHeight
+  [PrefKey.EditorFontSize]: EditorFontSize
 }