+++ packages/api/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.6](https://github.com/standardnotes/app/compare/@standardnotes/api@1.8.5...@standardnotes/api@1.8.6) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/api
+
+## [1.8.5](https://github.com/standardnotes/app/compare/@standardnotes/api@1.8.4...@standardnotes/api@1.8.5) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/api
+
 ## [1.8.4](https://github.com/standardnotes/app/compare/@standardnotes/api@1.8.3...@standardnotes/api@1.8.4) (2022-09-28)
 
 **Note:** Version bump only for package @standardnotes/api