+++ packages/files/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.8](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.7...@standardnotes/files@1.10.8) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/files
+
+## [1.10.7](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.6...@standardnotes/files@1.10.7) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/files
+
 ## [1.10.6](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.5...@standardnotes/files@1.10.6) (2022-09-28)
 
 **Note:** Version bump only for package @standardnotes/files