+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.37.11](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.10...@standardnotes/mobile@3.37.11) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.37.10](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.9...@standardnotes/mobile@3.37.10) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.37.9](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.8...@standardnotes/mobile@3.37.9) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.37.8](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.7...@standardnotes/mobile@3.37.8) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.37.7](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.6...@standardnotes/mobile@3.37.7) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.37.6](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.5...@standardnotes/mobile@3.37.6) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.37.5](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.4...@standardnotes/mobile@3.37.5) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.37.4](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.3...@standardnotes/mobile@3.37.4) (2022-09-28)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.37.3](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.37.2...@standardnotes/mobile@3.37.3) (2022-09-28)
 
 **Note:** Version bump only for package @standardnotes/mobile