+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.102](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.101...@standardnotes/releases@1.3.102) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.101](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.100...@standardnotes/releases@1.3.101) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.100](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.99...@standardnotes/releases@1.3.100) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.99](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.98...@standardnotes/releases@1.3.99) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.98](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.97...@standardnotes/releases@1.3.98) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.97](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.96...@standardnotes/releases@1.3.97) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.96](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.95...@standardnotes/releases@1.3.96) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.95](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.94...@standardnotes/releases@1.3.95) (2022-09-28)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.3.94](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.93...@standardnotes/releases@1.3.94) (2022-09-28)
 
 **Note:** Version bump only for package @standardnotes/releases