+++ packages/ui-services/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.8](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.3.7...@standardnotes/ui-services@1.3.8) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
+## [1.3.7](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.3.6...@standardnotes/ui-services@1.3.7) (2022-09-29)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
 ## [1.3.6](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.3.5...@standardnotes/ui-services@1.3.6) (2022-09-28)
 
 **Note:** Version bump only for package @standardnotes/ui-services